change ns
