+++ android/app/src/main/cpp/native-lib.cpp
@@ -1,13 +1,7 @@
 #include <jni.h>
 #include <string>
 
-extern "C"
-JNIEXPORT jstring
-
-JNICALL
-Java_com_example_android_bazel_MainActivity_stringFromJNI(
-        JNIEnv *env,
-        jobject /* this */) {
-    std::string hello = "Hello from C++";
-    return env->NewStringUTF(hello.c_str());
+extern "C" JNIEXPORT jstring JNICALL Java_dev_pscm_android_MainActivity_stringFromJNI(JNIEnv *env, jobject /* this */) {
+  std::string hello = "Hello from C++";
+  return env->NewStringUTF(hello.c_str());
 }
