+++ android/app/src/main/java/dev/pscm/android/MainActivity.java
@@ -1,0 +1,28 @@
+package dev.pscm.android;
+
+import android.os.Bundle;
+import android.widget.TextView;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class MainActivity extends AppCompatActivity {
+
+  static {
+    System.loadLibrary("app");
+  }
+
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    setContentView(R.layout.activity_main);
+
+    // Example of a call to a native method
+    TextView tv = (TextView)findViewById(R.id.sample_text);
+    tv.setText(stringFromJNI());
+  }
+
+  /**
+   * A native method that is implemented by the 'native-lib' native library,
+   * which is packaged with this application.
+   */
+  public native String stringFromJNI();
+}
