+++ android/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.android.bazel.MainActivity"
+    tools:context="dev.pscm.android.MainActivity"
     >
 
     <TextView
